rename odt_rgbMonitor transforms to clarify intended behavior of digital cinema simulation
The device specified in this transform has a D65 white point. The transform does not include a chromatic adaptation step, which means that neutral ACES values will produce unequal display code values. The intent is to produce a D60 white on the D65 device to simulate the white point of the digital cinema display devices.

Also updates the images readme to reflect changes to naming of rgbMonitor odt.