refines highlight handling for "dcsim" ODTs to avoid undesirable clipping behavior
This adjustment is to correct an issue that exists in ODTs where the device is calibrated to a white chromaticity other than D60. In order to simulate D60 on such devices, unequal code values are sent to the display to achieve neutrals at D60.
For example, in order to produce D60 on a device calibrated to the DCI white point (i.e. equal code values yield CIE x,y chromaticities of 0.314, 0.351) the red channel is higher than green and blue to compensate for the "greenish" DCI white. This is the correct behavior but it means that as highlight increase, the red channel will hit the device maximum first and clip, resulting in a chromaticity shift as the green and blue channels continue to increase.

To avoid this clipping error, a slight scale factor is applied. This is enough to allow for the Rec709, Rec2020, and rgbMonitor ODTs to simulate D60 within their D65 calibration white point. However, the magnitude of the scale factor required for the P3DCI version was considered too large. Therefore, the required scale factor was reduced and the additional required compression was achieved via a reshaping of the highlight rolloff in conjunction with the scale. The shape of this rolloff was determined throught subjective experiments and deemed to best reproduce the "character" of the highlights in the P3D60 ODT.