fix erroneous luminance scaling found in hdr odts (issue #40)
HDR ODTs were improperly using a luminance-to-CV scaling with CINEMA_BLACK and CINEMA_WHITE parameters. These parameters are intended for scaling the "cinema" tonescale to code values and should not be used in the HDR ODTs, which instead use PQ to encode luminance as CVs.
The previously included Y_2_linCV scaling step was a copy/paste bug carried over from the regular ODTs and has been removed in this commit.

close issue #40