reverse changes made in commit 1b5a1e0ed remove extra lines of code in comments segmented_spline_c5_fwd
Cleaning up code for consistency and clarity. Duplicate endpoint coefficients had been added as a temporary fix for an occasional bug that was halting CTL processing using the transforms during v1.0 testing. For programmers trying to port the B-spline code to their own implementations, the duplicate coefficients and misalignment of the number of coefficients with the N_KNOTS value could be confusing. (Usually the number of coeffienients = the number of knots + 1)

With this correction, SegmentedSplineParams_c5 and SegmentedSplineParams_c9 will now have coefficient arrays with dimensions consistent to their name, taking five and nine coeffiecients, respectively.

In addition, there were some extra lines of code in the commented sections of segmented_spline_c5_fwd, which are provided for compatibility with CTL 1.4. If uncommented and used as they were, segmented_spline_c5_fwd could throw an "error: index out of range".