fixes minor math issues
* round() did not handle negative values correctly in all cases, since floor() rounds towards -infinity, instead of towards 0.
The fix removes the use of the floor function. Note that the function is not currently used in any ACES transforms, so this fixes the reliability of round() if it is used in the future, but does not affect any current ACES transforms.

* calculate_cat_matrix() performs unnecessary matrix transposes
Fix removes unnecessary matrix transposes in calculate_cat_matrix(). The math now matches standard form (e.g. http://www.brucelindbloom.com/index.html?Eqn_ChromAdapt.html)

Fix by contributor: selfshadow