fixes bug in piecewise function logic for lin_2_acesProxy()
The fix to the logic in the piecewise function makes the CTL more aligned with S-2013-001.
First, the cutoff value is a constant (equal to 2^(((CVmin-MidCVoffset)/StepsPerStop)-MidLogOffset) ) and should not be dependent on the "in" value to lin_to_ACESproxy.
Secondly, the specification returns CVmin for all ACESproxyLin "less than or equal to" 2^-9.72, whereas here it was incorrectly written as only "less than".