redefine CINEMA_BLACK to circumvent precision errors affecting intended system design
In the 48-nit ODTs, the intended return value for small ACES values is 0.02. However, due to the tonescale being applied in log space and the contants that define it being described in linear space, some precision errors can occur. The minimum return value from the tonescale is intended to be equal to 0.02, but in reality is equal to pow10(log10(0.02)). In order to get a CV of zero out of the ODT, it is important that CINEMA_BLACK and the tonescale minimum are exactly equal. Rather than redesign the tonescale, CINEMA_BLACK is redefined to exactly match the result of the tonescale, which is really really close to 0.02 and achieves the intended result. This minor change allows for the ODTs to return a CV of 0, whereas previously CV intended to be zero were actually 9.9037e-05 (normalized 0-1).