adjust HDR ODT tonescales to allow for code value of 0
This change changes the ODT min point mapping, increases the slope of the extension below the ODT min point, adjusts the 4 lowest coefficients, and subtracts a small offset. The combination of these effects allows for a CV of 0 to come out of the ST2048 encoding, which previously was limited to a 12-bit CV of 62 due to the 0.005 nit value assigned to the tonecurve's low map point. This does not dramatically change the normal part of the imagery but helps remove a haze floor that could be observed on monitors that can keep blacks below 0.005 nits.