revert function parameters of ODTs from 'uniform bool' to 'varying int'
This does not change the intent of the transforms. It simply changes the syntax and logic used to trigger the SMPTE legal range code block. There have been repeated reports of the CTL interpreter not correctly setting the default value to false when using a 'uniform bool'. The correct place to fix this would be in the CTL codebase, but I don't have resources to fix that. Therefore, this change back to the old logic that apparently works should allow users to utilize these transforms in instances where the CTL interpreter is behaving inconsistently.