add gamut compression transform
This transform is the result of work by the Gamut Compression Virtual Working Group. It is intended to address the same issues as the existing “Blue Highlight Fix” LMT but uses a more robust approach in an attempt to fix problematic colors while leaving the bulk of image values unchanged.