change logic for determining when to apply white scaling to a bool
This allows the user to control when the white is or isn't applied, and is clearer and more robust than relying on a mismatch in the white point chromaticities. For example, a DCDM P3D65 ODT already has white point fitting handled by a 48/52.37 scale factor, so wouldn't need that additional white scaling that would otherwise be applied because the encoding white and limiting white would differ.

This commit relies on a corresponding change to the Output Transform to add a "scale_white" bool and include it in the call to the display_encoding/decoding functions.